refactor(context,shared): split precedent corpus and env accessors
Split context/precedent.py (397L) into precedent_fingerprint.py (exact-match
fingerprinting), precedent_store.py (vector-store plumbing and backfill),
and precedent_corpus.py (semantic matcher and loader); precedent.py is now
a facade re-exporting every name. load_precedent_corpus's outer try/except
(fail-soft, never raises) stays intact, wrapping the newly extracted
_resolve_available_embedder/_embed_and_build_corpus helper calls. Test
patch targets on _resolve_vector_store and _BACKFILL_CHUNK keep working via
a deferred facade lookup (from quodeq.context import precedent as _facade),
matching the pattern already used in _cli_worktree.py.

Split shared/_env.py (336L) into _env_ai.py, _env_paths.py, _env_db.py, and
_env_embeddings.py; _env.py keeps the numeric helpers and ports/keys/urls
accessors and re-exports every moved name so the project-wide fan-in
(dozens of call sites) is unaffected.

Baseline shrinks 149 -> 142 entries (3 stale removed, no new violations).
