+++ src/quodeq/context/precedent_fingerprint.py
@@ -1,0 +1,100 @@
+"""Exact-match precedent fingerprinting.
+
+A precedent is a finding that was previously dismissed for this project.
+On the next evaluation, the scanner will likely surface the same code
+pattern again; without precedent tracking, the user has to dismiss it
+every run. This module computes a stable fingerprint for each dismissed
+finding so the post-LLM pipeline can downweight matches.
+
+Fingerprint = sha256 of ``(req, normalized_snippet)``. Whitespace and
+trailing punctuation are normalized so cosmetic edits to surrounding
+code don't break the match. Code identifiers are *not* normalized:
+renaming a variable produces legitimately different code.
+"""
+from __future__ import annotations
+
+import hashlib
+import re
+from pathlib import Path
+
+from quodeq.data.ports.precedents import DismissedSnippetsReader
+
+_WS_RE = re.compile(r"\s+")
+
+
+def _normalize_snippet(snippet: str | None) -> str:
+    """Collapse runs of whitespace and trim trailing punctuation/space."""
+    if not snippet:
+        return ""
+    collapsed = _WS_RE.sub(" ", snippet).strip()
+    return collapsed.rstrip(",;.")
+
+
+def fingerprint(req: str | None, snippet: str | None) -> str | None:
+    """Hex sha256 of ``req + '|' + normalized_snippet``, or None when blank.
+
+    Returning None for blank inputs lets callers skip lookup entirely
+    instead of poisoning the precedent set with a useless all-empty key.
+    """
+    norm = _normalize_snippet(snippet)
+    req_part = (req or "").strip()
+    if not req_part and not norm:
+        return None
+    payload = f"{req_part}|{norm}".encode("utf-8")
+    return hashlib.sha256(payload).hexdigest()
+
+
+def precedent_text(req: str | None, snippet: str | None) -> str | None:
+    """Canonical text embedded for a finding -- used on BOTH store and match
+    sides so the comparison is symmetric. None when both parts are blank.
+
+    Note: ``.rstrip()`` on top of ``_normalize_snippet`` mops up the trailing
+    space that punctuation-stripping can leave behind (e.g. "x = 1 ;" ->
+    "x = 1 "); it's applied here rather than in ``_normalize_snippet``
+    itself so ``fingerprint()``'s hash stays byte-for-byte unchanged.
+    """
+    norm = _normalize_snippet(snippet).rstrip()
+    req_part = (req or "").strip()
+    if not req_part and not norm:
+        return None
+    return f"{req_part}\n\n{norm}"
+
+
+def load_precedent_fingerprints(
+    project_dir: Path, *, read_dismissed: DismissedSnippetsReader | None = None,
+) -> set[str]:
+    """Load fingerprints for every dismissed finding in *project_dir*.
+
+    Aggregates across ``<run_id>/evaluation.db`` under *project_dir*. Missing
+    or locked DBs are skipped -- precedent matching degrades gracefully and
+    never breaks a scan.
+
+    *read_dismissed* is a test seam (see ``data/ports/precedents.py``);
+    production callers should inject
+    ``quodeq.data.sqlite.findings_queries.read_dismissed_snippets``
+    explicitly. Left unset, it's resolved lazily below -- the local import
+    avoids a cycle (``data.fs.repo_clone`` -> ``context`` -> here ->
+    ``data.sqlite`` would close a loop if this were a top-level import).
+
+    Legacy note: prior to PR 1 (live-grades), dismissals were stored in
+    ``<project_dir>/dismissed.json``. The migration in
+    ``data/migrations/dismissed_json_to_actions_log.py`` folds those legacy
+    entries into ``actions.jsonl`` on first projection, so once a project has
+    been opened post-deploy the SQL rows also capture the historical data.
+    """
+    if not project_dir or not project_dir.is_dir():
+        return set()
+
+    if read_dismissed is None:
+        from quodeq.data.sqlite.findings_queries import read_dismissed_snippets  # noqa: PLC0415
+        read_dismissed = read_dismissed_snippets
+
+    out: set[str] = set()
+    for run_dir in project_dir.iterdir():
+        if not run_dir.is_dir():
+            continue
+        for req, snippet in read_dismissed(run_dir):
+            fp = fingerprint(req, snippet)
+            if fp is not None:
+                out.add(fp)
+    return out
